chore: added useSafariFix hook
